each contract and actor needs a class to handle in and outs
