+++ aiken-contracts/fractional/scripts/batcher_v3/env.json
@@ -1,7 +1,7 @@
 {
     "sale_address": "addr_test1xqled05n6w7xrv7tgfvll642up2548a5g4wx98lralqs8aylpzd0j9ps8lwt0fdgn7mccd5q7etgyq6endsz4vqa6qqsxrjgm9",
     "queue_address": "addr_test1xzvs7sk6cpcy4fha2hruezzw2kxy8p0wg8xx7hd6pdryz7ulpzd0j9ps8lwt0fdgn7mccd5q7etgyq6endsz4vqa6qqs9y4j8d",
-    "order_book_address": "",
+    "order_book_address": "addr_test1xq9hduw3rta7klwemrdc57r6efmdg8xlpa6qclkqjlfq46ylpzd0j9ps8lwt0fdgn7mccd5q7etgyq6endsz4vqa6qqsf0wtlm",
     "batcher_address": "addr_test1vrs4fk7ea6rg2fvd00sa8um5unp0rt474kngwpc38v2z9vqujprdk",
     "batcher_pkh": "e154dbd9ee8685258d7be1d3f374e4c2f1aebeada68707113b1422b0",
     "collat_pkh": "b834fb41c45bd80e5fd9d99119723637fe9d1e3fc467bc1c57ae9aee",
@@ -11,7 +11,7 @@
     "network": "--testnet-magic 1",
     "socket_path": "/home/logic/Documents/Work/LogicalMechanism/full-node-wallet/node/db-testnet/node.socket",
     "data_ref_utxo": "ee06e854fc5e5616cdd231796e363cc78663dd6c1e7a4c7f48aebfa4734b6bdb#0",
-    "sale_ref_utxo": "f7e8c856656922e250c5f7c00d5dc82753a4398cbb54ac5c0c997a542a4cc477#1",
-    "queue_ref_utxo": "feb0b03a8820c46f8dd89a57ae02022401a1a1af2556b828b2171cfcc39862c5#1",
-    "order_book_ref_utxo": ""
+    "sale_ref_utxo": "c7948f5c376d2c68a3f2b34d1c6033b3e665808cd4a98c4b229c992f69a1bfaf#1",
+    "queue_ref_utxo": "6d7d4b4cf5b5bc076b0a5acf287c787387037d7d2edeacefd5dcc1e14c0ed9dd#1",
+    "order_book_ref_utxo": "616fe3296871332aad8df86588230a06ecceb21e09a72d5e067d3bf8e279b77b#1"
 }
