+++ aiken-contracts/fractional/scripts/data/queue/queue-datum.json
@@ -19,12 +19,12 @@
           "bytes": "2aa69f11649eff76766d9bff42e1ec0876c4deed8d8f1e53ef208542"
         },
         {
-          "bytes": "001bc280016af3f8db00b96af73fd2954bdb1363146c46c2b2980795d93ef43a"
+          "bytes": "001bc28001e9904e5ab98faacfe9f8cabf37404de9062ed424a22fc2796086b3"
         }
       ]
     },
     {
-      "int": 3
+      "int": 5
     },
     {
       "constructor": 0,
@@ -41,7 +41,7 @@
       ]
     },
     {
-      "bytes": "0022bfb0016af3f8db00b96af73fd2954bdb1363146c46c2b2980795d93ef43a"
+      "bytes": "0022bfb001e9904e5ab98faacfe9f8cabf37404de9062ed424a22fc2796086b3"
     }
   ]
 }
