+++ aiken-contracts/fractional/scripts/data/sale/sale-datum.json
@@ -19,7 +19,7 @@
           "bytes": "2aa69f11649eff76766d9bff42e1ec0876c4deed8d8f1e53ef208542"
         },
         {
-          "bytes": "001bc280016af3f8db00b96af73fd2954bdb1363146c46c2b2980795d93ef43a"
+          "bytes": "001bc28001e9904e5ab98faacfe9f8cabf37404de9062ed424a22fc2796086b3"
         },
         {
           "int": 390625
